Add resourceParser test.
